Program to print the first 50 fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	long long first = 1, second, temp = 0;
+	unsigned int first = 1, second, temp = 0;
 	int count;
 
 	for (count = 0; count < 50; count++)
@@ -17,7 +17,7 @@ int main(void)
 		second = temp + first;
 		temp = first;
 		first = second;
-		printf("%lld", second);
+		printf("%u", second);
 		if (count != 49)
 			printf(", ");
 	}
